Implement d128! as a procedural macro
This makes macro invocations free at runtime, whereas the old macro
delegated to from_str at runtime which has small but nonzero cost.

The old decimal::d128 macro is changed to be internal-only so the new
one doesn't have a conflicting name: the type and macro share a name so
it's not possible to import only one of them from a given crate.

The concept is based on https://github.com/alkis/decimal/pull/16
but completely rewritten with the newer (stable) procedural macros which
implies a requirement for rustc 1.31.0 (2018 edition). The
implementation uses the 2021 edition (rustc 1.56.0) for convenience.